refactor(mobile): decouple asset stores from screen modules
