+++ apps/mobile/src/screens/Home/utils/token.ts
@@ -10,17 +10,7 @@ import { makeSWRKeyAsyncFunc } from '@/core/utils/concurrency';
 import { BalanceEntity } from '@/databases/entities/balance';
 import type { ITokenItem } from '@/types/assets';
 
-export const queryTokensCache = makeSWRKeyAsyncFunc(
-  (user_id: string, isTestnet: boolean = false) => {
-    return requestOpenApiWithChainId(
-      ({ openapi }) => openapi.getCachedTokenList(user_id),
-      {
-        isTestnet,
-      },
-    );
-  },
-  ctx => [`${ctx.args[0]}-${ctx.args[1]}`],
-);
+export { queryTokensCache } from '@/core/apis/tokenCache';
 
 export const batchQueryTokens = makeSWRKeyAsyncFunc(
   async (
