+++ apps/mobile/src/store/nfts.ts
@@ -2,7 +2,7 @@ import { getTop10MyAccounts } from '@/core/apis/account';
 import { zCreate } from '@/core/utils/reexports';
 import { syncNFTs } from '@/databases/hooks/assets';
 import { NFTItemEntity } from '@/databases/entities/nftItem';
-import { DisplayNftItem } from '@/screens/Home/types';
+import type { DisplayNftItem } from '@/types/assets';
 import { eventBus, EventBusListeners } from '@/utils/events';
 import { useCallback, useEffect } from 'react';
 import { CollectionList } from '@rabby-wallet/rabby-api/dist/types';
