+++ apps/mobile/src/utils/requestQueue.ts
@@ -1,0 +1,7 @@
+import PQueue from 'p-queue';
+
+export const pQueue = new PQueue({
+  interval: 1000,
+  intervalCap: 100,
+  concurrency: 40,
+});
